Implements Scheduled Images Extension
The extension creates a new resource under servers/id to get, create
and delete it's scheduled image.

On the servers list and detail calls, it adds the system metadata
related to scheduled image service on the server dict in the response
object. It also enables filtering by the system metadata added by the
scheduled images service and if it exists, adds it on the servers
dictonaries.

bp https://blueprints.launchpad.net/nova/+spec/scheduled-images